refactor InstitutionsHttpApiSpec to match the refactor in InstitutionsHttpApi, where the file is broken into several pieces to easy the ability to work with it
