+++ api/src/test/scala/hmda/api/http/InstitutionSpec.scala
@@ -1,0 +1,40 @@
+package hmda.api.http
+
+import java.io.File
+
+import akka.http.scaladsl.model.{ ContentTypes, HttpEntity, Multipart }
+import akka.http.scaladsl.testkit.ScalatestRouteTest
+import com.typesafe.config.ConfigFactory
+import hmda.api.RequestHeaderUtils
+import hmda.persistence.HmdaSupervisor
+import hmda.persistence.HmdaSupervisor.FindActorByName
+import hmda.persistence.demo.DemoData
+import hmda.persistence.institutions.InstitutionPersistence
+import org.iq80.leveldb.util.FileUtils
+import org.scalatest.{ BeforeAndAfterAll, MustMatchers, WordSpec }
+
+trait InstitutionSpec extends WordSpec with MustMatchers with BeforeAndAfterAll
+    with ScalatestRouteTest with RequestHeaderUtils with InstitutionsHttpApi {
+
+  override def beforeAll(): Unit = {
+    val supervisor = HmdaSupervisor.createSupervisor(system)
+    supervisor ! FindActorByName(InstitutionPersistence.name)
+    DemoData.loadTestData(system)
+    super.beforeAll()
+  }
+
+  override def afterAll(): Unit = {
+    system.terminate()
+    val config = ConfigFactory.load()
+    val snapshotStore = new File(config.getString("akka.persistence.snapshot-store.local.dir"))
+    FileUtils.deleteRecursively(snapshotStore)
+  }
+
+  def multiPartFile(contents: String, fileName: String) =
+    Multipart.FormData(Multipart.FormData.BodyPart.Strict(
+      "file",
+      HttpEntity(ContentTypes.`text/plain(UTF-8)`, contents),
+      Map("filename" -> fileName)
+    ))
+
+}
